+++ packages/logger_crispin_transport/example/main.dart
@@ -3,7 +3,8 @@ import 'package:logger_crispin_transport/logger_crispin_transport.dart';
 
 void main() {
   Crispin().addTransport(LoggerCrispinTransport(LoggerCrispinTransportOptions(level: 'warn')));
-  Crispin().warn('this is a warning', meta: StackTrace.current.toString());
+  Crispin().warn('this is a warning', stackTrace: StackTrace.current);
+  Crispin().warn('this is a warning without StackTrace');
   Crispin().info('Not enabled');
   try {
     throw Exception('Error for example');