+++ packages/crispin/CHANGELOG.md
@@ -1,9 +1,16 @@
-## 0.1.0
+## 1.1.0
+- Add `removeTransportByType` allowing removal of transports by a `Type`
+- Add `removeAllTransports`
 
-- Core Crispin proxy including methods `info`, `warn`, `error`, `debug`, and `silly`
-- Ability to add a Transport (Crispin created Transports coming soon...)
+## 1.0.0
+- All log level methods now accept optional `stackTrace` parameter
 
 ## 0.2.0
 
 - dart formatting resolved
 - documentation including support for docs.page
+
+## 0.1.0
+
+- Core Crispin proxy including methods `info`, `warn`, `error`, `debug`, and `silly`
+- Ability to add Transports via `addTransport(..)`